+++ crates/wasmi/src/instance.rs
@@ -10,7 +10,10 @@ use super::{
     Stored,
     Table,
 };
-use crate::module::{FuncIdx, ModuleImportType};
+use crate::{
+    module::{FuncIdx, ModuleImportType},
+    ExternType,
+};
 use alloc::{
     boxed::Box,
     collections::{btree_map, BTreeMap},
@@ -111,27 +114,85 @@ impl InstanceEntity {
     }
 }
 
-/// An iterator over the [`Extern`] declarations of an [`Instance`].
-#[derive(Debug)]
-pub struct ExportsIter<'a> {
-    iter: btree_map::Iter<'a, Box<str>, Extern>,
+/// An exported WebAssembly value.
+///
+/// This type is primarily accessed from the [`Instance::exports`] method
+/// and describes what names and items are exported from a Wasm [`Instance`].
+#[derive(Debug, Clone)]
+pub struct Export<'instance> {
+    /// The name of the exported item.
+    name: &'instance str,
+    /// The definition of the exported item.
+    definition: Extern,
 }
 
-impl<'a> ExportsIter<'a> {
+impl<'instance> Export<'instance> {
+    /// Creates a new [`Export`] with the given `name` and `definition`.
+    pub(crate) fn new(name: &'instance str, definition: Extern) -> Export<'instance> {
+        Self { name, definition }
+    }
+
+    /// Returns the name by which this export is known.
+    pub fn name(&self) -> &'instance str {
+        self.name
+    }
+
+    /// Return the [`ExternType`] of this export.
+    ///
+    /// # Panics
+    ///
+    /// If `ctx` does not own this [`Export`].
+    pub fn ty(&self, ctx: impl AsContext) -> ExternType {
+        self.definition.ty(ctx)
+    }
+
+    /// Consume this [`Export`] and return the underlying [`Extern`].
+    pub fn into_extern(self) -> Extern {
+        self.definition
+    }
+
+    /// Returns the underlying [`Func`], if the [`Export`] is a function or `None` otherwise.
+    pub fn into_func(self) -> Option<Func> {
+        self.definition.into_func()
+    }
+
+    /// Returns the underlying [`Table`], if the [`Export`] is a table or `None` otherwise.
+    pub fn into_table(self) -> Option<Table> {
+        self.definition.into_table()
+    }
+
+    /// Returns the underlying [`Memory`], if the [`Export`] is a linear memory or `None` otherwise.
+    pub fn into_memory(self) -> Option<Memory> {
+        self.definition.into_memory()
+    }
+
+    /// Returns the underlying [`Global`], if the [`Export`] is a global variable or `None` otherwise.
+    pub fn into_global(self) -> Option<Global> {
+        self.definition.into_global()
+    }
+}
+
+/// An iterator over the [`Extern`] declarations of an [`Instance`].
+#[derive(Debug)]
+pub struct ExportsIter<'instance> {
+    iter: btree_map::Iter<'instance, Box<str>, Extern>,
+}
+
+impl<'instance> ExportsIter<'instance> {
     /// Creates a new [`ExportsIter`].
-    fn new(iter: btree_map::Iter<'a, Box<str>, Extern>) -> Self {
+    fn new(iter: btree_map::Iter<'instance, Box<str>, Extern>) -> Self {
         Self { iter }
     }
 
     /// Prepares an item to match the expected iterator `Item` signature.
     #[allow(clippy::borrowed_box)]
-    fn convert_item((name, export): (&'a Box<str>, &'a Extern)) -> (&'a str, &'a Extern) {
-        (&**name, export)
+    fn convert_item((name, export): (&'instance Box<str>, &'instance Extern)) -> Export {
+        Export::new(&**name, *export)
     }
 }
 
-impl<'a> Iterator for ExportsIter<'a> {
-    type Item = (&'a str, &'a Extern);
+impl<'instance> Iterator for ExportsIter<'instance> {
+    type Item = Export<'instance>;
 
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.iter.size_hint()